Move hash_array function to popgen_numba_fns
