+++ sgkit/utils.py
@@ -5,8 +5,6 @@ from typing import Any, Callable, Hashable, List, Mapping, Optional, Set, Tuple,
 import numpy as np
 from xarray import Dataset
 
-from sgkit.accelerate import numba_guvectorize
-
 from . import variables
 from .typing import ArrayLike, DType
 
@@ -393,37 +391,6 @@ def max_str_len(a: ArrayLike) -> ArrayLike:
         return lens
     else:
         return lens.max()
-
-
-@numba_guvectorize(  # type: ignore
-    [
-        "void(int8[:], int64[:])",
-        "void(int16[:], int64[:])",
-        "void(int32[:], int64[:])",
-        "void(int64[:], int64[:])",
-    ],
-    "(n)->()",
-)
-def hash_array(x: ArrayLike, out: ArrayLike) -> None:  # pragma: no cover
-    """Hash entries of ``x`` using the DJBX33A hash function.
-
-    This is ~5 times faster than calling ``tobytes()`` followed
-    by ``hash()`` on array columns. This function also does not
-    hold the GIL, making it suitable for use with the Dask
-    threaded scheduler.
-
-    Parameters
-    ----------
-    x
-        1D array of type integer.
-
-    Returns
-    -------
-    Array containing a single hash value of type int64.
-    """
-    out[0] = 5381
-    for i in range(x.shape[0]):
-        out[0] = out[0] * 33 + x[i]
 
 
 def smallest_numpy_int_dtype(value: int) -> Optional[DType]:
